Add files via upload
All record files was downloaded to our computers. we have to combine them all at once using this code and define it to one specific folder in our computers. With that said, our combined file will generated as XML file and the jobs already done. the RAW Data can be added to Excel Workbook and we can completely do some analyzing data.